fix firmware clang-tidy gate
Teach the documented clang-tidy invocation how to parse the ARM/newlib build and add narrowly justified suppressions for Pico fixed-address access, bounded snprintf on newlib, legacy serial parser conversions, similar typed firmware APIs, generated headers, and legacy runtime globals.

Resolve the remaining actionable tidy findings with behavior-preserving cleanup: explicit string comparisons, non-reserved locals, wider loop counters, explicit float math casts, a small sync warning helper, and an impossible-path reload lane guard.

Generated-By: Codex (GPT-5)

+++ firmware/src/tmc2209.c
@@ -119,7 +119,7 @@ uint8_t tmc_crc8(const uint8_t *data, uint8_t len) {
     uint8_t crc = 0;
     for (uint8_t i = 0; i < len; i++) {
         uint8_t b = data[i];
-        for (uint8_t j = 0; j < TMC_CRC_BITS_PER_BYTE; j++) {
+        for (int j = 0; j < TMC_CRC_BITS_PER_BYTE; j++) {
             uint8_t mix = (crc >> TMC_CRC_TOP_BIT_SHIFT) ^ (b & 1u);
             crc <<= 1;
             if (mix) {
